Do not signal sender to abort if a retrieve conn exits.(#17343)
In the previous design, before a retrieve connection exits, it would
cancel all partially retrieved endpoints in this session, which could
lead to the entire transaction being aborted. However, in some
scenarios, partially retrieving data and then exiting is considered
reasonable behavior. For example, if I only want to fetch 100 records
(as specified by the SQL 'LIMIT' clause), it is acceptable to exit after
enough data has been retrieved without triggering any errors. Therefore,
in this commit, we use SIGUSR1 instead of SIGINT to notify the sender,
allowing it to return to the DoingCommandRead state.